fancy up tools/add_copyright.py, and add more copyrights

+++ tools/add_copyright.py
@@ -5,10 +5,12 @@
 
 '''Automatically add tiresome licensing boilerplate.'''
 
-import os, sys
+import os, sys, datetime, re
 
-copyright = '''\
-Copyright 2013 Evan Laforge
+dry_run = True
+
+copyright_template = '''\
+Copyright %s Evan Laforge
 This program is distributed under the terms of the GNU General Public
 License 3.0, see COPYING or http://www.gnu.org/licenses/gpl-3.0.txt
 '''
@@ -18,34 +20,42 @@ def main():
         process(fn)
 
 def process(fn):
+    template = add_comments(fn, copyright_template)
+    copyright = template % datetime.datetime.now().year
+    copyright_pattern = re.escape(template).replace('\\%s', '201[0-9]')
+
+    with open(fn) as fp:
+        contents = fp.read()
+
+    if contents.startswith('#!'):
+        (shebang, contents) = contents.split('\n', 1)
+    else:
+        shebang = ''
+    if re.match(copyright_pattern, contents):
+        # print fn, 'already has copyright'
+        return
+    print 'adding copyright to', fn
+    if dry_run:
+        return
+    tmp = fn + '.tmp'
+    with open(tmp, 'w') as fp:
+        fp.write(shebang)
+        fp.write(copyright)
+        if not contents.startswith('\n'):
+            fp.write('\n')
+        fp.write(contents)
+    os.chmod(tmp, os.stat(fn).st_mode)
+    os.rename(tmp, fn)
+
+def add_comments(fn, txt):
     if any(fn.endswith(e) for e in ('.hs', '.hsc')):
         comment = '--'
     elif any(fn.endswith(e) for e in ('.cc', '.c', '.h')):
         comment = '//'
     else:
         comment = '#'
-    msg = ''.join(
-        '%s %s\n' % (comment, line) for line in copyright.split('\n')
-        if line)
-    with open(fn) as fp:
-        line = fp.readline()
-        shebang = ''
-        if line.startswith('#!'):
-            shebang = line
-            line = fp.readline()
-        if line == msg[:msg.index('\n')+1]:
-            return
-        contents = line + fp.read()
-    print 'add copyright to', fn
-    tmp = fn + '.tmp'
-    with open(tmp, 'w') as fp:
-        fp.write(shebang)
-        fp.write(msg)
-        if not contents.startswith('\n'):
-            fp.write('\n')
-        fp.write(contents)
-    os.chmod(tmp, os.stat(fn).st_mode)
-    os.rename(tmp, fn)
+    return ''.join(
+        '%s %s\n' % (comment, line) for line in txt.split('\n') if line)
 
 if __name__ == '__main__':
     main()
